Backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
newspaper3k==0.2.8
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
# Use lxml-binary for Windows compatibility, fallback to html.parser
lxml>=4.9.0; platform_system != "Windows"
nltk==3.8.1
feedparser==6.0.10
schedule==1.2.0
# Additional Windows-friendly dependencies
html5lib==1.1
cssselect==1.3.0
